requirements.txt
langchain>=1.2.0
langchain-google-genai>=2.0.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-classic>=0.1.0
langgraph>=0.2.0
ddgs>=9.0.0
pydantic>=2.12.0
pydantic-settings>=2.12.0
faiss-cpu>=1.8.0
rank_bm25>=0.2.2
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
langfuse>=3.0.0
deepeval>=1.6.0
pytest>=8.0.0
